Product Model Changed

+++ lib/features/home/controller/get_product_types_controller.dart
@@ -21,7 +21,7 @@ class GetProductTypesController extends GetxController {
       },
       (data) async {
         productTypes.value = data;
-        selectedProductType.value = productTypes.value?.data[0].id ?? 0;
+        selectedProductType.value = productTypes.value?.data[0].id.toInt() ?? 0;
         await Get.find<GetProductsByTypeController>().getProductsByType(
           typeID: selectedProductType.value,
         );
